BZ1310208: When converting .XLS to GDT, Globals defined with short Class Names are converted to invalid Globals file.
(cherry picked from commit 5e335d848d72f2d2fe1a4c84dcb89eb1a8742687)
