+++ drools-wb-screens/drools-wb-dtable-xls-editor/drools-wb-dtable-xls-editor-backend/src/test/resources/logback-test.xml
@@ -16,6 +16,16 @@
   -->
 <configuration>
 
+    <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
+      <encoder>
+        <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+      </encoder>
+    </appender>
+
+    <root level="error">
+      <appender-ref ref="STDOUT" />
+    </root>
+
   <logger name="org.drools" level="info"/>
 
 </configuration>
